Generator + GateOR
Added removing elements in Generator
Added GateXOR
Repaired BoundryConection Checker

+++ WireworldinJava/src/Wireworld/generator/CellMouseListener.java
@@ -21,11 +21,11 @@ public class CellMouseListener implements MouseListener {
 
     @Override
     public void mouseClicked(MouseEvent me) {
-        /* if (tools.isDeleteOperation()) {
+        if (tools.isDeleteOperation()) {
             tools.deleteElement((JLabel) me.getComponent());
-        } else {*/
-        tools.changeElement((JLabel) me.getComponent(), true);
-        //}
+        } else {
+            tools.changeElement((JLabel) me.getComponent(), true);
+        }
     }
 
     @Override
@@ -39,16 +39,13 @@ public class CellMouseListener implements MouseListener {
 
     @Override
     public void mouseEntered(MouseEvent me) {
-        tools.changeElement((JLabel) me.getComponent(), false);
-        String name = ((JLabel) me.getComponent()).getName();
-        String loc[] = name.split("x");
-        int x = Integer.parseInt(loc[0]);
-        int y = Integer.parseInt(loc[1]);
-        States s = WireWorldManager.getInstance().getBoard().getPointOnBoard(x, y);
-        setComunicat(s.getElementNumber() + "", true);
+        if (!tools.isDeleteOperation()) {
+            tools.changeElement((JLabel) me.getComponent(), false);
+        }
     }
 
     @Override
-    public void mouseExited(MouseEvent me) {
+    public void mouseExited(MouseEvent me
+    ) {
     }
 }
